Fix bugs for new tf version

+++ object_detector_retinanet/keras_retinanet/initializers.py
@@ -34,6 +34,6 @@ class PriorProbability(keras.initializers.Initializer):
 
     def __call__(self, shape, dtype=None):
         # set bias to -log((1 - p)/p) for foreground
-        result = np.ones(shape, dtype=dtype) * -math.log((1 - self.probability) / self.probability)
+        result = np.ones(shape, dtype=np.float32) * -math.log((1 - self.probability) / self.probability)
 
         return result
